+++ CHANGELOG.md
@@ -10,6 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - :boom: Merged `useInternalFormContext` and `react-hook-form context` in the same `useFormContext` context
+- 
+## [1.6.0] - 2023-09-25
+
+### Added
+
+- Added support to have undefined values in a select input by using a magic guid value
+- `datePickerRef` ref prop to `DatePickerInput`
+
 
 ## [1.5.0] - 2023-09-18
 
@@ -306,7 +314,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Created package :tada:
 
-[unreleased]: https://github.com/neolution-ch/react-hook-form-components/compare/1.5.0...HEAD
+[unreleased]: https://github.com/neolution-ch/react-hook-form-components/compare/1.6.0...HEAD
 [0.1.2]: https://github.com/neolution-ch/react-hook-form-components/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/neolution-ch/react-hook-form-components/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/neolution-ch/react-hook-form-components/releases/tag/0.1.0
@@ -331,6 +339,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.4.0]: https://github.com/neolution-ch/react-hook-form-components/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/neolution-ch/react-hook-form-components/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/neolution-ch/react-hook-form-components/releases/tag/0.2.0
+[1.6.0]: https://github.com/neolution-ch/react-hook-form-components/compare/1.5.0...1.6.0
 [1.5.0]: https://github.com/neolution-ch/react-hook-form-components/compare/1.4.0...1.5.0
 [1.4.0]: https://github.com/neolution-ch/react-hook-form-components/compare/1.3.3...1.4.0
 [1.3.3]: https://github.com/neolution-ch/react-hook-form-components/compare/1.3.2...1.3.3